refactor(models): move LocaleData to dedicated model file
Restructure codebase by moving LocaleData class from locale_constants.dart to a dedicated model file. This improves code organization and follows better separation of concerns.
