+++ lib/utils/imports/common_libs.dart
@@ -54,6 +54,12 @@ export '../../route/routes.dart';
 export 'package:montreal/route/router.dart';
 
 // =============================================================================
+// MODELS
+// =============================================================================
+
+export '../../model/locale_data.dart';
+
+// =============================================================================
 // CONSTANTS
 // =============================================================================
